Quick restructuring LibraryScreen.kt
